@@ -2,41 +2,60 @@ package command
 
 import (
 	"bytes"
-	"text/template"
-
+	"context"
+	"github.com/k0sproject/rig/v2/cmd"
+	"github.com/sikalabs/gobble/pkg/host"
 	"github.com/sikalabs/gobble/pkg/libtask"
-	"github.com/sikalabs/gobble/pkg/utils/exec_utils"
+	"github.com/sikalabs/gobble/pkg/logger"
+	"github.com/sikalabs/gobble/pkg/utils"
+	"text/template"
+	"time"
 )
 
-type TaskCommand struct {
-	Cmd string `yaml:"cmd"`
+type Task struct {
+	libtask.BaseTask        // Embed BaseTask
+	Cmd              string `yaml:"cmd"`
 }
 
-func Run(
-	taskInput libtask.TaskInput,
-	taskParams TaskCommand,
-) libtask.TaskOutput {
+func (t *Task) Run(taskInput libtask.TaskInput, host *host.Host) libtask.TaskOutput {
+	// Create a new context with a timeout
+	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
+	defer cancel()
 
-	tmpl, err := template.New("cmd").Parse(taskParams.Cmd)
+	// Parse the command template
+	tmpl, err := template.New("cmd").Parse(t.Cmd)
 	if err != nil {
 		return libtask.TaskOutput{
 			Error: err,
 		}
 	}
+	// Execute the command template
 	var buf bytes.Buffer
 	err = tmpl.Execute(&buf, map[string]interface{}{
 		"Config": taskInput.Config,
-		"Vars":   taskInput.Vars,
+		"Vars":   utils.MergeMaps(taskInput.Vars, host.Vars),
 	})
 	if err != nil {
 		return libtask.TaskOutput{
 			Error: err,
 		}
 	}
-	err = exec_utils.SSH(
-		taskInput,
-		buf.String(),
-	)
+
+	// Determine the client to use
+	client := host.Client
+	if taskInput.Sudo {
+		client = host.Client.Sudo()
+	}
+
+	// Execute the command
+	err = client.ExecContext(ctx, buf.String(),
+		cmd.LogInput(true),
+		cmd.StreamOutput(),
+		cmd.LogError(true),
+		cmd.Logger(logger.Slog))
+	if err != nil {
+		logger.Log.Warnf("host: '%s' task: '%s' failed", host.Client.Address(), t.GetName())
+	}
 	return libtask.TaskOutput{
 		Error: err,
 	}
