@@ -2,10 +2,10 @@ package yaml_utils
 
 import (
 	"fmt"
-	"log"
+	"github.com/sikalabs/gobble/pkg/logger"
 	"os"
 
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 func MergeYAMLs(paths []string) {
@@ -16,20 +16,20 @@ func MergeYAMLs(paths []string) {
 	for _, path := range paths {
 		file, err := os.ReadFile(string(path))
 		if err != nil {
-			log.Fatalf("Failed to read %s: %v", string(path), err)
+			logger.Log.Fatalf("Failed to read %s: %v", string(path), err)
 		}
 		data = append(data, file...)
 	}
 
 	err := yaml.Unmarshal(data, &merged)
 	if err != nil {
-		log.Fatalf("Failed to merge YAML files: %v", err)
+		logger.Log.Fatalf("Failed to merge YAML files: %v", err)
 	}
 
 	// Print the merged YAML
 	out, err := yaml.Marshal(merged)
 	if err != nil {
-		log.Fatalf("Failed to marshal merged YAML: %v", err)
+		logger.Log.Fatalf("Failed to marshal merged YAML: %v", err)
 	}
 	fmt.Println(string(out))
 }
